feat: associated todos with user entity

+++ src/todo/todo.service.ts
@@ -4,6 +4,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { Todo } from './entities/todo.entity';
+import { User } from 'src/users/entities/user.entity';
 
 @Injectable()
 export class TodoService {
@@ -11,27 +12,35 @@ export class TodoService {
     @InjectRepository(Todo) private todoRepository: Repository<Todo>,
   ) {}
 
-  create(createTodoDto: CreateTodoDto): Promise<Todo> {
-    return this.todoRepository.save(this.todoRepository.create(createTodoDto));
+  create(createTodoDto: CreateTodoDto, userId: User['id']): Promise<Todo> {
+    return this.todoRepository.save(
+      this.todoRepository.create({ ...createTodoDto, user: { id: userId } }),
+    );
   }
 
-  findAll(): Promise<Todo[]> {
-    return this.todoRepository.find();
+  findAll(userId: User['id']): Promise<Todo[]> {
+    return this.todoRepository.find({ where: { user: { id: userId } } });
   }
 
-  async findOne(id: number): Promise<Todo> {
-    const todo: Todo = await this.todoRepository.findOneBy({ id });
+  async findOne(id: number, userId: User['id']): Promise<Todo> {
+    const todo: Todo = await this.todoRepository.findOneBy({
+      id,
+      user: { id: userId },
+    });
     if (!todo) throw new NotFoundException('Todo not found');
     return todo;
   }
 
-  async update(id: number, updateTodoDto: UpdateTodoDto) {
-    await this.findOne(id);
-    return this.todoRepository.update({ id }, updateTodoDto);
+  async update(id: number, updateTodoDto: UpdateTodoDto, userId: User['id']) {
+    await this.findOne(id, userId);
+    return this.todoRepository.update(
+      { id, user: { id: userId } },
+      updateTodoDto,
+    );
   }
 
-  async remove(id: number) {
-    await this.findOne(id);
-    return this.todoRepository.delete({ id });
+  async remove(id: number, userId: User['id']) {
+    await this.findOne(id, userId);
+    return this.todoRepository.delete({ id, user: { id: userId } });
   }
 }
